[cleanup] moves ProgressBar.js to containers, cleans up server.js code
