+++ client/containers/ProgressBar.js
